Added automatic upload to TikTok
Used the tiktok_uploader module to automatically upload the generated video to TikTok. This only seems to work sometimes, possibly because there needs to be some time between uploads. In addition, I made it ignore reddit posts with links in them to avoid having the TTS read out the entire link