Check for incorrect timepoints in API response
If the timepoints in the API response are in the wrong order, the program will now print an error and also add the post's ID to the database to avoid processing it again. Another column was added to the database to track whether each video was actually uploaded. Lastly, the text processing was updated to better account for multiple punctuation marks at the end of a sentence.