Added logging and improved error handling
Added a log function to write messages to a log file. Also added a fatal_error function that logs and prints the reason for the error, adds the current post to the database if needed, commits and closes the database, and ends the program. Lastly, added a catch for exceptions that could be thrown by the Google API call.